build-aux: fix warning in the built _get_string() methods with -Wsign-compare
  mbim-error-types.c: In function ‘mbim_core_error_get_string’:
  mbim-error-types.c:53:17: warning: comparison of integer expressions of different signedness: ‘MbimCoreError’ {aka ‘enum <anonymous>’} and ‘gint’ {aka ‘const int’} [-Wsign-compare]
     53 |         if (val == mbim_core_error_values[i].value)
        |                 ^~

  mbim-enum-types.c: In function ‘mbim_cid_basic_connect_get_string’:
  mbim-enum-types.c:68:17: warning: comparison of integer expressions of different signedness: ‘MbimCidBasicConnect’ {aka ‘enum <anonymous>’} and ‘gint’ {aka ‘const int’} [-Wsign-compare]
     68 |         if (val == mbim_cid_basic_connect_values[i].value)
        |                 ^~

  mbim-enum-types.c: In function ‘mbim_data_class_build_string_from_mask’:
  mbim-enum-types.c:1958:18: warning: comparison of integer expressions of different signedness: ‘MbimDataClass’ {aka ‘enum <anonymous>’} and ‘guint’ {aka ‘const unsigned int’} [-Wsign-compare]
   1958 |         if (mask == mbim_data_class_values[i].value) {
        |                  ^~
    CC       libmbim_glib_generated_la-mbim-basic-connect.lo

+++ build-aux/templates/mbim-enum-types-template.c
@@ -51,7 +51,7 @@ const gchar *
     guint i;
 
     for (i = 0; @enum_name@_values[i].value_nick; i++) {
-        if (val == @enum_name@_values[i].value)
+        if ((gint)val == @enum_name@_values[i].value)
             return @enum_name@_values[i].value_nick;
     }
 
@@ -82,7 +82,7 @@ gchar *
 
     for (i = 0; @enum_name@_values[i].value_nick; i++) {
         /* We also look for exact matches */
-        if (mask == @enum_name@_values[i].value) {
+        if ((guint)mask == @enum_name@_values[i].value) {
             if (str)
                 g_string_free (str, TRUE);
             return g_strdup (@enum_name@_values[i].value_nick);
